+++ core/src/main/java/de/bitbrain/braingdx/input/UpdateableInputMultiplexer.java
@@ -1,0 +1,17 @@
+package de.bitbrain.braingdx.input;
+
+import com.badlogic.gdx.InputMultiplexer;
+import com.badlogic.gdx.InputProcessor;
+import de.bitbrain.braingdx.util.Updateable;
+
+public class UpdateableInputMultiplexer extends InputMultiplexer implements Updateable {
+
+   @Override
+   public void update(float delta) {
+      for (InputProcessor processor : getProcessors()) {
+         if (processor instanceof Updateable) {
+            ((Updateable) processor).update(delta);
+         }
+      }
+   }
+}