+++ core/src/main/java/de/bitbrain/braingdx/util/Updateable.java
@@ -1,0 +1,9 @@
+package de.bitbrain.braingdx.util;
+
+/**
+ * Indicates that the implementation of this interface is updateable per frame.
+ */
+public interface Updateable {
+
+   void update(float delta);
+}