+++ core/src/test/java/de/bitbrain/braingdx/event/GameEventManagerTest.java
@@ -2,6 +2,7 @@ package de.bitbrain.braingdx.event;
 
 import com.badlogic.gdx.Application;
 import com.badlogic.gdx.Gdx;
+import de.bitbrain.braingdx.util.GdxUtils;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -21,7 +22,7 @@ public class GameEventManagerTest {
    @Before
    public void beforeTest() {
       impl = new GameEventManagerImpl();
-      Gdx.app = mock(Application.class);
+      GdxUtils.mockApplicationContext();
       event = new ConcreteGameEvent();
    }